refactor(core): group process and terminal concerns into core/runtime/
Phase 3 of giving src/core an interior. terminal, pager, jobControl,
shutdown, projectRoot, appStateFile, updateNotice, and startupNotice
move from core root into core/runtime/. Pure move with no symbol
renames; all eight have outside consumers, so the module declares no
interior. core root now holds only types.ts, reviewDigest.ts, and
liveComments.ts beside the module directories.
