+++ docs/changelog-on-hunk-dev.md
@@ -106,6 +106,6 @@ path list, or release commits will not trigger a deploy.
   launch-video pipeline keeps generated media out of Git.
 - **Contributor lists.** The GitHub release bodies name first-time contributors, which is community
   goodwill and organic links. `CHANGELOG.md` does not carry authors, so this needs a second input.
-- **The in-app update notice.** `src/core/updateNotice.ts` tells users a new version exists without
+- **The in-app update notice.** `src/core/runtime/updateNotice.ts` tells users a new version exists without
   linking what changed. Appending `hunk.dev/changelog/<minor>` is the highest-intent entry point
   available and is deliberately left as a separate change.
