+++ docs/source-architecture.md
@@ -15,6 +15,9 @@ src/core/changeset/  the changeset model and the pipeline that acquires one: loa
                      per-file construction, sidecar/source reads, and hunk formatting
 src/core/invocation/ how a run is asked for: command inputs, layered configuration, the
                      command catalog, user-facing errors, paths, and version
+src/core/runtime/    the process and terminal a run lives in: TTY capabilities, the pager,
+                     job control, shutdown, project-root discovery, persisted app state,
+                     and startup/update notices
 src/core/theme/      bundled theme metadata, custom-theme rules, and terminal theme detection
 src/core/watch/      input signatures, observation plans/backends, and refresh coordination
 src/core/vcs/        provider-neutral VCS catalog, contracts, operation dispatch, and host support
