+++ src/app/sessionSelector.ts
@@ -1,4 +1,4 @@
-import { findProjectRootCandidate } from "../core/projectRoot";
+import { findProjectRootCandidate } from "../core/runtime/projectRoot";
 import type { SessionSelectorInput } from "../core/types";
 import type { VcsCatalog } from "../core/vcs/types";
