+++ src/core/runtime/startupNotice.ts
