+++ src/extensions/apply.ts
@@ -2,7 +2,7 @@ import {
   BUILT_IN_FILE_LANGUAGE_EXTENSIONS,
   registerFileLanguage,
 } from "../core/changeset/fileLanguage";
-import type { StartupNotice } from "../core/startupNotice";
+import type { StartupNotice } from "../core/runtime/startupNotice";
 import type { Changeset } from "../core/types";
 import { detectVcs, extendVcsCatalog, getDefaultVcsAdapter } from "../core/vcs";
 import type { VcsAdapter, VcsCatalog } from "../core/vcs/types";
