+++ src/extensions/host.ts
@@ -1,5 +1,5 @@
 import { pathToFileURL } from "node:url";
-import { findProjectRootCandidate } from "../core/projectRoot";
+import { findProjectRootCandidate } from "../core/runtime/projectRoot";
 import { EXTENSION_ID_RULE, HUNK_VENDOR_EXTENSION_ID, isValidExtensionId } from "./extensionIds";
 import { bindExtensionEventBus } from "./events";
 import { registerHostRuntimeModules } from "./hostRuntimeModules";
